Moving the folder helpers to the correct position and updating paths
