+++ client/index.js
@@ -5,5 +5,5 @@
 import {AppRegistry} from 'react-native';
 import App from './js/App.js';
 import {name as appName} from './app.json';
-
+console.disableYellowBox = true;
 AppRegistry.registerComponent(appName, () => App);
