+++ client/js/lib/helpers/interest_function.js
