--- client/js/screens/UserProfile/helpers/source.js
@@ -1,20 +1,0 @@
-let source = {
-  coffee: {
-    true: require(`../../../assets/PNG/ways_to_meet/coffee_active.png`),
-    false: require(`../../../assets/PNG/ways_to_meet/coffee_inactive.png`),
-  },
-  after_school: {
-    true: require(`../../../assets/PNG/ways_to_meet/after_school_active.png`),
-    false: require(`../../../assets/PNG/ways_to_meet/after_school_inactive.png`),
-  },
-  lunch: {
-    true: require(`../../../assets/PNG/ways_to_meet/lunch_active.png`),
-    false: require(`../../../assets/PNG/ways_to_meet/lunch_inactive.png`),
-  },
-  walk: {
-    true: require(`../../../assets/PNG/ways_to_meet/walk_active.png`),
-    false: require(`../../../assets/PNG/ways_to_meet/walk_inactive.png`),
-  },
-};
-
-export default source;
